+++ crates/zng-wgt-inspector/src/live/inspector_window.rs
@@ -602,6 +602,7 @@ fn info_watchers(wgt: &InspectedWidget) -> UiNode {
         let flash = value_background(&value);
         children.push(Text! {
             txt = value;
+            paragraph_indent = (1.em(), true); // hang next lines
             font_color = PROPERTY_VALUE_COLOR_VAR;
             background_color = flash;
         });